Conversion Log File (example)
Demonstrates logging and workflow of the script. Two new recordings get converted, one of which one has a special subtitle stream that is specifically mapped by the script rules.
One old files gets deleted in the end.
This is the actual logfile, created by the script along with the execution.