Update to version 1.2.4
Found out that WMC:s smaller title field is in fact used in some Finnish shows to include the name of and episode in some series.
To see this text in Kodi (without altering the skin files) I needed to an engine, that adds the "sub-title" in plot field, before the actual plot/description. Similar to channel adding, only this defaults to empty no data is found (either no lines with fingerprint or the resulting data is blank. This works differently with channel names, where there will be a vertical bar | as a placeholder even if the channel name is not found.