Updated to BACAR v.1.5 with tools etc
Since I moved to using TVHeadend on a virtual machine there was a need for some big additions. The most notable is the ability to choose from three different tagging methods (to tag the source file as converted), as renaming wasn't transparent enough.
To make the setup easier I wrote some tools that can do something that would otherwise take ages. They are mostly needed in setting up the script and especially if something goes wrong.