Option to allow multiple streams per rule.
There are occasions, where two (or more) streams can not be differentiated based on ffprobes simple reporting (for example two subtitle streams with same codec and same language). So this adds the variable "allowmultiplestreamhits" to each rule, so that all of the found streams will be mapped. This option also makes the script more ignorant about insufficient rules, so be careful when using this. Only apply it to situations, where you know that the rules work but the files have "double streams". The problem was first seen with Finnish Yle HD channel recordings, that have a two Finnish subtitle streams (the other is for audio impaired, but ffprobe sees no difference in them).
This is the only change from v.1.5 to v.1.5.1.